Ankush/gmx fix (#592)
* gmx feedback

* small bug fix

* feedback changes

* feedback changes

* refcatored disconnect feature and added auto close

* auto modal open
