+++ packages/notifi-dapp-example/src/app/notifi/layout.tsx
@@ -1,11 +1,10 @@
 'use client';
 
-import { Icon } from '@/assets/Icon';
+import Disconnect from '@/components/Disconnect';
 import { NotifiContextWrapper } from '@/context/NotifiContextWrapper';
-import { isEVMChain } from '@/utils/typeUtils';
 import { useWallets } from '@notifi-network/notifi-wallet-provider';
 import { usePathname, useRouter } from 'next/navigation';
-import React, { useEffect, useState } from 'react';
+import React, { useEffect } from 'react';
 
 export default function NotifiSingupLayout({
   children,
@@ -13,7 +12,6 @@ export default function NotifiSingupLayout({
   children: React.ReactNode;
 }) {
   const { wallets, selectedWallet } = useWallets();
-  const [isWalletMenuOpen, setIsWalletMenuOpen] = useState<boolean>(false);
   const pathname = usePathname();
   const router = useRouter();
 
@@ -23,51 +21,15 @@ export default function NotifiSingupLayout({
     }
   }, [selectedWallet]);
 
-  const keys = selectedWallet && wallets[selectedWallet].walletKeys;
-  const accountAddress = keys
-    ? isEVMChain(keys)
-      ? keys.hex?.toLowerCase()
-      : keys.bech32
-    : '';
-
-  const showDisconnectButton =
-    ['/notifi/ftu', '/notifi/signup'].includes(pathname) && accountAddress;
+  const showDisconnectButton = ['/notifi/ftu', '/notifi/signup'].includes(
+    pathname,
+  );
 
   return (
     <>
       {showDisconnectButton ? (
         <div className={`fixed z-40 top-7 right-0.5 hidden md:block`}>
-          <div className={`rounded-[0.75rem] overflow-hidden`}>
-            <div
-              className={`cursor-pointer flex px-4 text-notifi-tab-unselected-text w-52 py-2 ${
-                isWalletMenuOpen
-                  ? 'bg-notifi-wallet-menu-card-bg border border-transparent'
-                  : ''
-              } 
-              `}
-              onClick={() => setIsWalletMenuOpen(!isWalletMenuOpen)}
-            >
-              <Icon
-                id={'user-protrait'}
-                className={`text-notifi-icon-unselected mt-[0.125rem]`}
-              />
-              <div className="ml-2">
-                {accountAddress.slice(0, 6)} ... {accountAddress.slice(-6)}
-              </div>
-            </div>
-
-            {isWalletMenuOpen ? (
-              <div
-                className={`flex px-4 text-notifi-tab-unselected-text w-52 py-2 bg-notifi-wallet-menu-card-bg border-t border-[#565A8D]`}
-                onClick={() => {
-                  if (keys) wallets[selectedWallet].disconnect();
-                }}
-              >
-                <Icon id={'leave'} className={`text-notifi-icon-unselected`} />
-                <div className="ml-2 cursor-pointer">Disconnect</div>
-              </div>
-            ) : null}
-          </div>
+          <Disconnect />
         </div>
       ) : null}
       <NotifiContextWrapper>{children}</NotifiContextWrapper>;
